Implemented fixes suggested by Akash and Pantazis. All tests pass.
