+++ Source/Core/Timers/ProductionTimerMachine.cs
@@ -1,5 +1,5 @@
 ﻿//-----------------------------------------------------------------------
-// <copyright file="TimerModel.cs">
+// <copyright file="ProductionTimerMachine.cs">
 //      Copyright (c) Microsoft Corporation. All rights reserved.
 // 
 //      THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
@@ -12,94 +12,104 @@
 // </copyright>
 //-----------------------------------------------------------------------
 
-
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Timers;
 
 namespace Microsoft.PSharp.Timers
 {
 	/// <summary>
-	/// A timer model, used for testing purposes.
+	/// Wrapper class for a system timer.
 	/// </summary>
-    public class TimerModel : Machine
-    {
-		#region static fields
+	public class ProductionTimerMachine : Machine
+	{
+		#region fields
 
 		/// <summary>
-		/// Adjust the probability of firing a timeout event.
+		/// Specified if periodic timeout events are desired.
 		/// </summary>
-		public static int NumStepsToSkip = 1;
+		private bool IsPeriodic;
 
-		#endregion
-
-		#region private fields
+		/// <summary>
+		/// Specify the periodicity of timeout events.
+		/// </summary>
+		private int Period;
 
 		/// <summary>
 		/// Machine to which eTimeout events are dispatched.
 		/// </summary>
-		private MachineId client;
-
-		/// <summary>
-		/// True if periodic eTimeout events are desired.
-		/// </summary>
-		private bool IsPeriodic;
+		private MachineId Client;
 
         /// <summary>
         /// TimerId
         /// </summary>
         private TimerId tid;
 
-        #endregion
+		/// <summary>
+		/// System timer to generate Elapsed timeout events in production mode.
+		/// </summary>
+		private System.Timers.Timer timer;
 
-        #region states
-        [Start]
-		[OnEntry(nameof(InitializeTimer))]
-		[OnEventDoAction(typeof(HaltTimerEvent), nameof(DisposeTimer))]
-        [OnEventDoAction(typeof(RepeatTimeout), nameof(SendTimeout))]
-        private class Init : MachineState { }
+		/// <summary>
+		/// Flag to prevent timeout events being sent after stopping the timer.
+		/// </summary>
+		private bool IsTimerEnabled = false;
+
+		/// <summary>
+		/// Used to synchronize the Elapsed event handler with timer stoppage.
+		/// </summary>
+		private readonly Object tlock = new object();
 
 		#endregion
 
-		#region event handlers
+		#region states
+		[Start]
+		[OnEntry(nameof(InitializeTimer))]
+		[OnEventDoAction(typeof(HaltTimerEvent), nameof(DisposeTimer))]
+		private class Init : MachineState { }
+
+		#endregion
+
+		#region handlers
 		private void InitializeTimer()
 		{
 			InitTimer e = (this.ReceivedEvent as InitTimer);
-			this.client = e.client;
+			this.Client = e.client;
 			this.IsPeriodic = e.IsPeriodic;
+			this.Period = e.Period;
             this.tid = e.tid;
-            this.Send(this.Id, new RepeatTimeout());
-		}
 
-		private void SendTimeout()
-		{
-            this.Assert(NumStepsToSkip >= 0);
+			this.IsTimerEnabled = true;
+			this.timer = new System.Timers.Timer(Period);
 
-            // If not periodic, send a single timeout event
-            if (!this.IsPeriodic)
+			if (!IsPeriodic)
 			{
-                // Probability of firing timeout is atmost 1/N
-                if ((this.RandomInteger(NumStepsToSkip) == 0) && this.FairRandom())
-                {
-                    this.Send(this.client, new TimerElapsedEvent(tid));
-                }
-                else
-                {
-                    this.Send(this.Id, new RepeatTimeout());
-                }
-            }
-			else
-			{				
-				// Probability of firing timeout is atmost 1/N
-				if ((this.RandomInteger(NumStepsToSkip)==0) && this.FairRandom())
-				{
-				   this.Send(this.client, new TimerElapsedEvent(tid));
-				}
-				this.Send(this.Id, new RepeatTimeout());
+				this.timer.AutoReset = false;
 			}
-			
+
+			this.timer.Elapsed += ElapsedEventHandler;
+			this.timer.Start();
+		}
+		#endregion
+
+		#region private methods
+		/// <summary>
+		/// Handler for the Elapsed event generated by the system timer.
+		/// </summary>
+		/// <param name="source"></param>
+		/// <param name="e"></param>
+		private void ElapsedEventHandler(Object source, ElapsedEventArgs e)
+		{
+            lock (this.tlock)
+			{
+				if (this.IsTimerEnabled)
+				{
+					Runtime.SendEvent(this.Client, new TimerElapsedEvent(tid));
+				}
+			}
 		}
 
 		private void DisposeTimer()
@@ -107,18 +117,26 @@ namespace Microsoft.PSharp.Timers
 			HaltTimerEvent e = (this.ReceivedEvent as HaltTimerEvent);
 
 			// The client attempting to stop this timer must be the one who created it.
-			this.Assert(e.client == this.client);
+			this.Assert(e.client == this.Client);
+
+			lock (this.tlock)
+			{
+				this.IsTimerEnabled = false;
+				this.timer.Stop();
+				this.timer.Dispose();
+			}
 
 			// If the client wants to flush the inbox, send a markup event.
 			// This marks the endpoint of all timeout events sent by this machine.
 			if (e.flush)
 			{
-				this.Send(this.client, new Markup());
+				this.Send(this.Client, new Markup());
 			}
 
 			// Stop this machine
 			this.Raise(new Halt());
 		}
 		#endregion
+
 	}
 }
